refactor(renderer): restore HelpActions name
