--- packages/renderer/src/lib/help/HelpActionsItems.svelte
@@ -1,6 +1,0 @@
-<script lang="ts">
-import HelpActions from '/@/lib/help/HelpActions.svelte';
-import { Items } from '/@/lib/help/HelpItems';
-</script>
-
-<HelpActions items={Items}/>
